deleted class UtilsTest unused
